requirements.txt
attrs==19.3.0
Automat==0.8.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.5.0
parsel==1.5.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
queuelib==1.5.0
requests==2.22.0
Scrapy==1.8.0
service-identity==18.1.0
six==1.14.0
Twisted==19.10.0
urllib3==1.25.8
w3lib==1.21.0
zope.interface==4.7.1
